agent-backend/requirements.txt
fastapi
uvicorn
GitPython
python-dotenv
requests
